test-requirement.txt
odoo-addon-partner-invoicing-mode @ git+https://github.com/OCA/account-invoicing.git@refs/pull/1515/head#subdirectory=setup/partner_invoicing_mode
odoo-addon-partner-invoicing-mode-at-shipping @ git+https://github.com/OCA/account-invoicing.git@refs/pull/1628/head#subdirectory=setup/partner_invoicing_mode_at_shipping
odoo-addon-partner-invoicing-mode-fourteen-days @ git+https://github.com/OCA/account-invoicing.git@refs/pull/1534/head#subdirectory=setup/partner_invoicing_mode_fourteen_days
